requirements.txt
matplotlib==3.5.1
numpy==1.21.4
Pillow==8.4.0